+++ src/components/index.js
@@ -36,6 +36,9 @@ export Search from './theme/Search/Search';
 export SearchTags from './theme/Search/SearchTags';
 export TabularView from './theme/View/TabularView';
 export View from './theme/View/View';
+export Comments from './theme/Comments/Comments';
+export SocialSharing from './theme/SocialSharing/SocialSharing';
+export Tags from './theme/Tags/Tags';
 
 export Actions from './manage/Actions/Actions';
 export Add from './manage/Add/Add';